requirements.txt
Django==3.2.3
djangorestframework==3.12.4
python-dotenv==0.17.1
psycopg2-binary==2.8.6
requests==2.25.1
gunicorn==20.1.0